Godfile execution-engine step 3: extract task-runner-review-gate
Slice 3 of the execution-engine god-file decomposition. Moves the
review-gate/merge-gate polling family and CI-failure publication out of
task-runner.ts into task-runner-review-gate.ts as free functions over a
Pick<TaskRunner, ...> host. The ReviewGateCiFailure* declarations relocate
with their owning cluster and stay barrel-visible via index.ts, so app-layer
importers compile unchanged. TaskRunner keeps thin one-line delegates.
Behavior is intentionally identical.
