+++ packages/execution-engine/src/index.ts
@@ -30,6 +30,10 @@ export type {
   ExecutionPoolConfig,
   PoolSelection,
 } from './task-runner-pool.js';
+export type {
+  ReviewGateCiFailureTrigger,
+  ReviewGateCiFailureLifecyclePublisher,
+} from './task-runner-review-gate.js';
 export * from './merge-runner.js';
 export * from './conflict-resolver.js';
 export * from './merge-gate-provider.js';
